Resolves #1
