+++ bin/src/main/resources/build-info.properties
@@ -1,0 +1,3 @@
+# Contains informations aboud DPU build. This file is used by
+# cz.cuni.mff.xrg.uv.boost.dpu.vaadin.AboutTab. Do not modify it's content!
+build.timestamp = ${build.timestamp}
