+++ bin/src/main/resources/resources.properties
@@ -1,0 +1,28 @@
+dpu.name =  e-relationalFromCKAN
+dpu.name.menu = e-relationalFromCKAN
+
+# Dialog
+CkanVaadinDialog.columns.as.strings = Ignore Database column types (use Strings)
+CkanVaadinDialog.tablename.label = Table name:
+CkanVaadinDialog.filter.label = Filter resources:
+CkanVaadinDialog.logs.label = Logs:
+CkanVaadinDialog.logs.item.selection.changed = Selection changed to resource:
+CkanVaadinDialog.tree.button.expand.all = Expand all
+CkanVaadinDialog.tree.button.collapse.all = Collapse all
+CkanVaadinDialog.tree.item.tooltip.no_description = No description available
+CkanVaadinDialog.tree.label = Available relational dataset resources:
+CkanVaadinDialog.tree.loading.label = Loading datasets, please wait ...
+CkanVaadinDialog.tree.org.for.datasets.without.org.label = Datasets without organization  
+CkanVaadinDialog.selected.item.log = Selected item: {0}
+
+# Messages
+dpu.ckan.starting = {0} is starting.
+dpu.create.new.table = Creating new table
+dpu.created.new.table = New table was created
+
+# Errors
+errors.api.missing = Configuration error: Missing catalog API location URL
+errors.failed.retrieve.datasets = Failed to retrieve datasets from catalog.
+errors.failed.retrieve.datastore.resource.id.list = Failed to retrieve datastore resource list
+errors.failed.retrieve.org.datasets = Failed to retrieve private datasets from catalog for organization {0}.
+errors.token.missing = Configuration error: Missing authentication token for catalog API
