+++ bin/src/main/resources/resources_sk.properties
@@ -1,0 +1,28 @@
+dpu.name =  Stiahnutie d\u00e1t z katal\u00f3gu
+dpu.name.menu = d\u00e1ta z katal\u00f3gu
+
+# Dialog
+CkanVaadinDialog.columns.as.strings = Ignoruj typy databazovych st\u013apcov (pou\u017ei re\u0165azce)
+CkanVaadinDialog.tablename.label = N\u00e1zov tabu\u013eky:
+CkanVaadinDialog.filter.label = Filter zdrojov:
+CkanVaadinDialog.logs.label = Logy:
+CkanVaadinDialog.logs.item.selection.changed = V\u00fdber zmenen\u00fd na zdroj:
+CkanVaadinDialog.tree.button.expand.all = Rozbali\u0165 v\u0161etko
+CkanVaadinDialog.tree.button.collapse.all = Zbali\u0165 v\u0161etko
+CkanVaadinDialog.tree.item.tooltip.no_description = Popis nie je k dispoz\u00edcii
+CkanVaadinDialog.tree.label = Dostupn\u00e9 rela\u010dn\u00e9 zdroje datasetov:
+CkanVaadinDialog.tree.loading.label = Na\u010d\u00edtaj\u00fa sa datasety, pros\u00edm \u010dakajte ...
+CkanVaadinDialog.tree.org.for.datasets.without.org.label = Datasety bez organiz\u00e1cie 
+CkanVaadinDialog.selected.item.log = Vybran\u00e1 polo\u017eka: {0}
+
+# Messages
+dpu.ckan.starting = Sp\u00fa\u0161\u0165a sa {0}.
+dpu.create.new.table = Vytv\u00e1ra sa nov\u00e1 tabu\u013eka
+dpu.created.new.table = Bola vytvoren\u00e1 nov\u00e1 tabu\u013eka
+
+# Errors
+errors.api.missing = Chyba v nastaven\u00ed: Ch\u00fdba URL na API katal\u00f3gu.
+errors.failed.retrieve.datasets = Nepodarilo sa na\u010d\u00edta\u0165 datasety z katal\u00f3gu.
+errors.failed.retrieve.datastore.resource.id.list = Nepodarilo sa na\u010d\u00edta\u0165 zoznam rela\u010dn\u00fdch zdrojov z katal\u00f3gu.
+errors.failed.retrieve.org.datasets = Nepodarilo sa na\u010d\u00edta\u0165 s\u00fakromn\u00e9 datasety z katal\u00f3gu pre organiz\u00e1ciu {0}.
+errors.token.missing = Chyba v nastaven\u00ed: Ch\u00fdba autentifika\u010dn\u00fd token pre API katal\u00f3gu
