--- src/main/java/eu/unifiedviews/plugins/extractor/ckan/relational/CkanTreeItem.java
@@ -1,17 +1,0 @@
-package eu.unifiedviews.plugins.extractor.ckan.relational;
-
-public interface CkanTreeItem {
-
-    public String getId();
-    public String getName();
-    /**
-     * 
-     * @return String showed in tooltip of the tree item
-     */
-    public String getDescription();
-    /**
-     * 
-     * @return String showed as "name" of the tree item
-     */
-    public String toString();
-}
