--- src/main/java/eu/unifiedviews/plugins/extractor/ckan/relational/RelationalFromCkan.java
@@ -1,156 +1,0 @@
-package eu.unifiedviews.plugins.extractor.ckan.relational;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import eu.unifiedviews.dataunit.DataUnit;
-import eu.unifiedviews.dataunit.DataUnitException;
-import eu.unifiedviews.dataunit.relational.WritableRelationalDataUnit;
-import eu.unifiedviews.dpu.DPU;
-import eu.unifiedviews.dpu.DPUContext;
-import eu.unifiedviews.dpu.DPUContext.MessageType;
-import eu.unifiedviews.dpu.DPUException;
-import eu.unifiedviews.helpers.dataunit.resource.Resource;
-import eu.unifiedviews.helpers.dataunit.resource.ResourceHelpers;
-import eu.unifiedviews.helpers.dpu.config.ConfigHistory;
-import eu.unifiedviews.helpers.dpu.context.ContextUtils;
-import eu.unifiedviews.helpers.dpu.exec.AbstractDpu;
-
-
-@DPU.AsExtractor
-public class RelationalFromCkan extends AbstractDpu<RelationalFromCkanConfig_V1> {
-
-    private static final Logger LOG = LoggerFactory.getLogger(RelationalFromCkan.class);
-
-    public static final String CONFIGURATION_SECRET_TOKEN = "dpu.uv-e-relationalFromCKAN.secret.token";
-    public static final String CONFIGURATION_CATALOG_API_LOCATION = "dpu.uv-e-relationalFromCKAN.catalog.api.url";
-    
-    /**
-     * Requesting rows in chunks
-     */
-    public static final int REQUEST_RECORD_LIMIT = 1000;
-
-    @DataUnit.AsOutput(name = "output")
-    public WritableRelationalDataUnit relationalOutput;
-
-    private DPUContext context;
-    
-    public RelationalFromCkan() {
-        super(RelationalFromCkanVaadinDialog.class, ConfigHistory.noHistory(RelationalFromCkanConfig_V1.class));
-    }
-
-    @Override
-    protected void innerExecute() throws DPUException {
-        this.context = this.ctx.getExecMasterContext().getDpuContext();
-        String shortMessage = this.ctx.tr("dpu.ckan.starting", this.getClass().getSimpleName());
-        String longMessage = String.valueOf(this.config);
-        this.context.sendMessage(DPUContext.MessageType.INFO, shortMessage, longMessage);
-        
-        Map<String, String> environment = this.context.getEnvironment();
-        final long pipelineId = this.context.getPipelineId();
-        final String userId = this.context.getPipelineExecutionOwnerExternalId();
-        final String token = environment.get(CONFIGURATION_SECRET_TOKEN);
-        final String catalogApiLocation = environment.get(CONFIGURATION_CATALOG_API_LOCATION);
-        
-        if (token == null || token.isEmpty()) {
-            throw ContextUtils.dpuException(this.ctx, "errors.token.missing");
-        }
-        
-        if (catalogApiLocation == null || catalogApiLocation.isEmpty()) {
-            throw ContextUtils.dpuException(this.ctx, "errors.api.missing");
-        }
-        
-        CatalogApiConfig apiConfig = new CatalogApiConfig(catalogApiLocation, pipelineId, userId, token);
-        
-        if (ctx.canceled()) {
-            throw ContextUtils.dpuExceptionCancelled(ctx);
-        }
-        
-        final String packageId = config.getPackageId();
-        final String resourceId = config.getResourceId();
-        
-        try {
-            LOG.debug("Processing CKAN resource with id {} from package / dataset with id {}", resourceId, packageId);
-            // create table from user config
-            DatastoreSearchResult tableMetadata = RelationalFromCkanHelper.getDatastoreSearchResult(apiConfig, resourceId, 0, 0);
-            
-            String createTableQuery = prepareCreateTableQuery(config, tableMetadata);
-            LOG.debug("Table create query: {}", createTableQuery);
-            ContextUtils.sendMessage(ctx, DPUContext.MessageType.DEBUG, ctx.tr("dpu.create.new.table"), createTableQuery);
-            DatabaseHelper.executeUpdate(createTableQuery, relationalOutput);
-            ContextUtils.sendMessage(ctx, DPUContext.MessageType.DEBUG,
-                    ctx.tr("dpu.created.new.table"),
-                    DatabaseHelper.getTableMetadataInfo(config.getTableName(), relationalOutput));
-
-            // add metadata
-            final String tableName = config.getTableName().toUpperCase();
-            relationalOutput.addExistingDatabaseTable(tableName, tableName);
-
-            Resource resource = ResourceHelpers.getResource(relationalOutput, tableName);
-            Date now = new Date();
-            resource.setCreated(now);
-            resource.setLast_modified(now);
-            ResourceHelpers.setResource(relationalOutput, tableName, resource);
-
-            // get the data from CKAN in chunks
-            boolean recordsRemain = true;
-            int offset = 0;
-            while (!ctx.canceled() && recordsRemain) {
-                LOG.debug("requesting record from {0} to {1}", offset, (offset + REQUEST_RECORD_LIMIT - 1));
-                DatastoreSearchResult result = RelationalFromCkanHelper.getDatastoreSearchResult(apiConfig, resourceId, REQUEST_RECORD_LIMIT, offset);
-                String insertIntoQuery = prepareInsertIntoQuery(config, result);
-                LOG.debug("Insert into query: {}", insertIntoQuery);
-                DatabaseHelper.executeUpdate(insertIntoQuery, relationalOutput);
-                recordsRemain = offset + REQUEST_RECORD_LIMIT < result.total;
-                offset += result.records.size();
-            }
-        } catch (DataUnitException e) {
-            ContextUtils.dpuException(ctx, e, "Failed to process datastore resource.");
-        } catch (Exception e) {
-            String errMsg = "Failed to retrieve datastore resource data.";
-            LOG.error(errMsg, e);
-            this.context.sendMessage(MessageType.ERROR, errMsg, e.getMessage());
-        }
-    }
-    
-    protected static String prepareInsertIntoQuery(RelationalFromCkanConfig_V1 config, DatastoreSearchResult result) {
-        StringBuilder sb = new StringBuilder();
-        
-        List<String> columnNames = result.getFieldList();
-        List<String> values = new ArrayList<String>(result.records.size());
-        for (Record r : result.records) {
-            values.add(r.getSqlInsertValues(columnNames, config.isColumnsAsString()));
-        }
-        
-        sb.append(String.format("INSERT INTO %s VALUES\n%s",
-                config.getTableName(),
-                StringUtils.join(values, ",\n")));
-        
-        sb.append(";");
-        
-        return sb.toString();
-    }
-
-    protected static String prepareCreateTableQuery(RelationalFromCkanConfig_V1 config, DatastoreSearchResult tableMetadata) {
-        // inspired by TabularToRelational
-
-        StringBuilder sb = new StringBuilder();
-        sb.append(String.format("CREATE TABLE %s (", config.getTableName()));
-        for (Iterator<Field> it = tableMetadata.fields.iterator(); it.hasNext();) {
-            sb.append(it.next().getStringForCreateTable(config.isColumnsAsString()));
-            if (it.hasNext()) {
-                sb.append(", ");
-            }
-        }
-        sb.append(");");
-        return sb.toString();
-    }
-
-}
